feat: editとnewを統一
